refactor:#66 폼 공통화 및 모드에 따른 조건부 렌더링 추가
