--- src/components/features/auth-form/login-form.tsx
@@ -1,69 +1,0 @@
-'use client';
-
-import { useForm } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { signupSchema } from '@/lib/utils/validation/auth-validate';
-import AuthFormInput from '@/components/features/auth-form/auth-form-input';
-import AuthFormContainer from '@/components/features/auth-form/auth-form-container';
-import FormButton from '@/components/features/auth-form/form-button-box';
-import type { AuthInputProps } from '@/types/components/auth-form';
-import type { SignupFormData } from '@/lib/utils/validation/auth-validate';
-import { AUTH } from '@/constants/auth-form';
-
-type FormData = Omit<AuthInputProps, 'register' | 'trigger' | 'getValues'>;
-
-const LoginForm = () => {
-  const {
-    register,
-    formState: { errors, isSubmitting, isValid },
-    getValues,
-    trigger
-  } = useForm<SignupFormData>({
-    resolver: zodResolver(signupSchema),
-    defaultValues: {
-      email: '',
-      password: ''
-    },
-    mode: 'onBlur'
-  });
-
-  const loginFormData: FormData[] = [
-    {
-      placeholder: '이메일 입력',
-      type: 'email',
-      name: 'email',
-      errorMessage: errors.email?.message,
-      label: '이메일'
-    },
-    {
-      placeholder: '비밀번호 입력',
-      type: 'password',
-      name: 'password',
-      checkButton: false,
-      errorMessage: errors.password?.message,
-      label: '비밀번호'
-    }
-  ];
-
-  return (
-    <AuthFormContainer>
-      {loginFormData.map((data, index) => (
-        <AuthFormInput
-          key={index}
-          register={register}
-          trigger={trigger}
-          getValues={getValues}
-          placeholder={data.placeholder}
-          type={data.type}
-          name={data.name}
-          checkButton={data.checkButton}
-          errorMessage={data.errorMessage}
-          label={data.label}
-        />
-      ))}
-      <FormButton mode={AUTH.LOGIN} isSubmitting={isSubmitting} isValid={isValid} />
-    </AuthFormContainer>
-  );
-};
-
-export default LoginForm;
